refactor: Name인터페이스 삭제 및 리팩토링
