+++ src/main/java/wooteco/subway/line/domain/LineName.java
@@ -1,47 +1,48 @@
-package wooteco.subway.name.domain;
+package wooteco.subway.line.domain;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
 
-public class StationName implements Name {
-    private static final Pattern PATTERN = Pattern.compile("^[가-힣|0-9]*역$");
+public class LineName {
+    private static final Pattern PATTERN = Pattern.compile("^[가-힣|0-9]*선$");
     private String name;
 
-    public StationName() {
+    private LineName() {
     }
 
-    public StationName(final String name) {
+    public LineName(final String name) {
         validateName(name);
         this.name = name;
     }
 
+    public static LineName emptyName() {
+        return new LineName();
+    }
+
     public void validateName(final String name) {
         if (!PATTERN.matcher(name).matches()) {
-            throw new IllegalArgumentException("잘못된 역 이름입니다.");
+            throw new IllegalArgumentException("잘못된 노선 이름입니다.");
         }
     }
 
-    @Override
     public String name() {
         return name;
     }
 
-    @Override
     public boolean sameName(final String name) {
         return this.name.equals(name);
     }
 
-    @Override
-    public Name changeName(final String name) {
-        return new StationName(name);
+    public LineName changeName(final String name) {
+        return new LineName(name);
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        StationName that = (StationName) o;
-        return Objects.equals(name, that.name);
+        LineName lineName = (LineName) o;
+        return Objects.equals(name, lineName.name);
     }
 
     @Override
