--- src/main/java/wooteco/subway/name/domain/Name.java
@@ -1,23 +1,0 @@
-package wooteco.subway.name.domain;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-@JsonTypeInfo(
-        use = JsonTypeInfo.Id.NAME,
-        include = JsonTypeInfo.As.PROPERTY,
-        property = "type",
-        defaultImpl = StationName.class
-)
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = StationName.class, name = "stationName"),
-        @JsonSubTypes.Type(value = LineName.class, name = "lineName"),
-        @JsonSubTypes.Type(value = NullName.class, name = "nullName"),
-})
-public interface Name {
-    String name();
-
-    boolean sameName(final String name);
-
-    Name changeName(String name);
-}
